rename: JwtProvider.class -> JwtService.class
#15
